requirements.txt
aiofiles==24.1.0
aiogram==3.14.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
decorator==5.1.1
Django==5.1.2
django-cors-headers==4.6.0
django-crontab==0.7.1
django-environ==0.11.2
django-filter==24.3
django-js-asset==2.2.0
django-mptt==0.16.0
django-multiselectfield==0.1.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-recursive==0.4.0
drf-spectacular==0.27.1
drf-yasg==1.21.8
executing==2.1.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
ipython==8.29.0
jedi==0.19.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
magic-filter==1.0.12
matplotlib-inline==0.1.7
msgpack==1.1.0
multidict==6.1.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyTelegramBotAPI==4.23.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.7
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.1
service-identity==24.2.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
stack-data==0.6.3
telebot==0.0.5
traitlets==5.14.3
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
views==0.3
vine==5.1.0
wcwidth==0.2.13
yarl==1.17.1
zope.interface==7.1.1
